requirements.txt
requests==2.31.0
python-dotenv==1.0.1
walmart-marketplace-api==1.1.0